rename pymongo sub-package
