+++ scout/adapter/__init__.py
@@ -1,1 +1,2 @@
 from scout.adapter.mongoengine.mongo import MongoEngineAdapter
+from scout.adapter.mongo.base import MongoAdapter
