+++ scout/adapter/mongo/__init__.py
