+++ scout/adapter/mongo/case.py
