+++ scout/adapter/mongo/event.py
