+++ scout/adapter/mongo/institute.py
