+++ scout/adapter/mongo/panel.py
