+++ scout/adapter/mongo/query.py
