+++ scout/adapter/mongo/variant.py
